cli: Allow missing images to be skipped from `image mirror`
It is useful to be able to skip images that don't exist in some
cases, and instead print a warning. Add `--skip-missing` to the
image mirror command.
