+++ contrib/completions/bash/oc
@@ -10910,6 +10910,8 @@ _oc_image_mirror()
     local_nonpersistent_flags+=("--max-registry=")
     flags+=("--s3-source-bucket=")
     local_nonpersistent_flags+=("--s3-source-bucket=")
+    flags+=("--skip-missing")
+    local_nonpersistent_flags+=("--skip-missing")
     flags+=("--skip-mount")
     local_nonpersistent_flags+=("--skip-mount")
     flags+=("--skip-multiple-scopes")
